Now that the folder has more than just mst in it, rename to sparsetools, which should help with referencing it.
