+++ sklearn/utils/sparsetools/_traversal.pyx
